backend/python/requirements.txt
annotated-types==0.7.0
anyio==3.6.2
asgiref==3.4.1
bcrypt==4.0.1
cffi==1.17.1
click==8.0.4
colorama==0.4.5
cryptography==37.0.1
fastapi==0.75.2
h11==0.13.0
halo==0.0.31
idna==3.4
log-symbols==0.0.14
loguru==0.7.2
MarkupSafe==2.1.5
packaging==25.0
paramiko==3.4.1
passlib==1.7.4
pipdeptree==2.26.1
pycparser==2.21
pydantic==1.9.2
pydantic_core==2.20.1
PyJWT==2.4.0
PyMySQL==1.0.2
PyNaCl==1.5.0
python-dotenv==1.0.1
python-multipart==0.0.5
requests==2.6.0
six==1.17.0
sniffio==1.2.0
spinners==0.0.24
SQLAlchemy==1.4.36
starlette==0.17.1
termcolor==3.0.1
typing_extensions==4.12.2
uvicorn==0.17.6
Werkzeug==1.0.1
win32-setctime==1.1.0
xmltodict==0.13.0
